requirements.txt
colorama==0.4.6
iniconfig==2.3.0
packaging==25.0
pip==25.3
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.1
selene==2.0.0rc9
selenium==4.38.0
python-dotenv==1.2.1
isort==7.0.0
flake8==7.3.0
black==25.9.0

allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.4.0

